requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
bcrypt==4.1.2
fpdf2==2.7.7
gunicorn==21.2.0
python-dotenv==1.0.0
